pyproject.toml
[project]
name = "homework-2"
version = "0.1.0"
description = "Homework 2"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "dash>=4.0.0",
    "dash-core-components>=2.0.0",
    "dash-html-components>=2.0.0",
    "dash-table>=5.0.0",
    "huggingface-hub>=1.3.7",
    "jupyter>=1.1.1",
    "kagglehub>=0.4.2",
    "matplotlib>=3.10.8",
    "notebook>=7.5.3",
    "numpy>=2.4.2",
    "openai>=1.0.0",
    "pandas>=2.3.3",
    "penngrader-client>=0.5.13",
    "plotly[express]>=6.5.2",
    "polars>=1.37.1",
    "pydantic>=2.10.0",
    "python-dotenv>=1.2.1",
]
